compliance_bot/requirements.txt
streamlit
langchain
langchain-openai
langchain-community
langgraph
faiss-cpu
sentence-transformers
python-dotenv
langchain-aws
boto3
langchain-google-genai
langchain-huggingface
fastapi
uvicorn
python-multipart
